--- a/.gitignore
+++ b/.gitignore
@@ -24,6 +24,7 @@ wheels/
 MANIFEST
 __pycache__/
 *.pyc
+*.swp
 
 # Unit test / coverage reports
 htmlcov/