… python/vtext/tests/test_vectorize.py
… src/vectorize/tests.rs